refactor(agent-core): split loop facade
