+++ packages/agent-core/agent-core.cabal
@@ -91,7 +91,8 @@ library
                     , Agent.Tools.ShellReadOnly
                     , Agent.Tools.Types
                     , Agent.Transport.WebSocket
-    other-modules:    Paths_agent_core
+    other-modules:    Agent.Loop.Internal
+                    , Paths_agent_core
     autogen-modules:  Paths_agent_core
     hs-source-dirs:   src
     build-depends:    base >= 4.14 && < 5
